test(storage): move delete regression
