+++ src/dbus_server/item.rs
@@ -378,213 +378,3 @@ impl Item<'static> {
         Ok(time_to_int(metadata.modified()))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::{
-        collections::HashMap,
-        os::unix::fs::PermissionsExt,
-        path::{Path, PathBuf},
-        sync::Arc,
-    };
-
-    use redb::{
-        Database, MultimapTableDefinition, ReadableMultimapTable, ReadableTable, TableDefinition,
-    };
-    use tempfile::tempdir;
-    use tokio::{fs, net::UnixStream, process::Command};
-    use zbus::{connection::Builder, Guid};
-
-    use super::Item;
-    use crate::{
-        dbus_server::utils::secret_path,
-        pass::PasswordStore,
-        secret_store::{redb::RedbSecretStore, RedbHashMap, SecretStore, PASS_SUBDIR},
-    };
-
-    const ATTRIBUTES_TABLE: MultimapTableDefinition<(&str, &str), &str> =
-        MultimapTableDefinition::new("attributes");
-    const ATTRIBUTES_TABLE_REVERSE: TableDefinition<&str, RedbHashMap<&str, &str>> =
-        TableDefinition::new("attributes-reverse");
-    const LABELS_TABLE: TableDefinition<&str, &str> = TableDefinition::new("labels");
-
-    struct GpgAgentCleanup(PathBuf);
-
-    impl Drop for GpgAgentCleanup {
-        fn drop(&mut self) {
-            let _ = std::process::Command::new("gpgconf")
-                .arg("--homedir")
-                .arg(&self.0)
-                .args(["--kill", "gpg-agent"])
-                .status();
-        }
-    }
-
-    async fn init_gpg_home(gpg_home: &Path) -> String {
-        const RECIPIENT: &str = "pass-secret-service-test@example.invalid";
-
-        fs::create_dir(gpg_home).await.unwrap();
-        fs::set_permissions(gpg_home, std::fs::Permissions::from_mode(0o700))
-            .await
-            .unwrap();
-        let output = Command::new("gpg")
-            .arg("--batch")
-            .arg("--homedir")
-            .arg(gpg_home)
-            .args([
-                "--passphrase",
-                "",
-                "--quick-generate-key",
-                RECIPIENT,
-                "rsa2048",
-                "encr",
-                "0",
-            ])
-            .output()
-            .await
-            .unwrap();
-        assert!(
-            output.status.success(),
-            "failed to create test GPG key: {}",
-            String::from_utf8_lossy(&output.stderr)
-        );
-
-        RECIPIENT.to_owned()
-    }
-
-    #[tokio::test]
-    async fn item_delete_removes_all_secret_storage() {
-        let temp = tempdir().unwrap();
-        let gpg_home = temp.path().join("gnupg");
-        let password_store_dir = temp.path().join("password-store");
-        let _gpg_agent_cleanup = GpgAgentCleanup(gpg_home.clone());
-        let recipient = init_gpg_home(&gpg_home).await;
-        fs::create_dir(&password_store_dir).await.unwrap();
-        fs::write(password_store_dir.join(".gpg-id"), recipient)
-            .await
-            .unwrap();
-
-        let pass = Box::leak(Box::new(PasswordStore::for_test(
-            password_store_dir.clone(),
-            &gpg_home,
-        )));
-        let store = RedbSecretStore::new(pass).await.unwrap();
-        let collection_id = Arc::new(
-            store
-                .create_collection(Some("test".into()), None)
-                .await
-                .unwrap(),
-        );
-        let attributes = HashMap::from([
-            ("application".to_owned(), "regression-test".to_owned()),
-            ("account".to_owned(), "example".to_owned()),
-        ]);
-        let secret_id = Arc::new(
-            store
-                .create_secret(
-                    collection_id.clone(),
-                    Some("Secret label".into()),
-                    b"secret value".to_vec(),
-                    Arc::new(attributes.clone()),
-                )
-                .await
-                .unwrap(),
-        );
-        let secret_file = password_store_dir
-            .join(PASS_SUBDIR)
-            .join(&*collection_id)
-            .join(format!("{secret_id}.gpg"));
-
-        assert!(secret_file.exists());
-        assert_eq!(
-            store
-                .get_secret_label(collection_id.clone(), secret_id.clone())
-                .await
-                .unwrap(),
-            "Secret label"
-        );
-        assert_eq!(
-            store
-                .read_secret_attrs(collection_id.clone(), secret_id.clone())
-                .await
-                .unwrap(),
-            attributes
-        );
-        for (key, value) in &attributes {
-            let matching_ids = store
-                .search_collection(
-                    collection_id.clone(),
-                    Arc::new(HashMap::from([(key.clone(), value.clone())])),
-                )
-                .await
-                .unwrap();
-            assert!(matching_ids.contains(secret_id.as_ref()));
-        }
-
-        let guid = Guid::generate();
-        let (client_stream, server_stream) = UnixStream::pair().unwrap();
-        let (client, server) = futures_util::try_join!(
-            Builder::unix_stream(client_stream).p2p().build(),
-            Builder::unix_stream(server_stream)
-                .server(guid)
-                .unwrap()
-                .p2p()
-                .build(),
-        )
-        .unwrap();
-        let item_path = secret_path(&*collection_id, &*secret_id).unwrap();
-        let item = Item {
-            collection_id: collection_id.clone(),
-            id: secret_id.clone(),
-            store: Box::new(store.clone()),
-            last_access: Default::default(),
-            notify_on_access: false,
-        };
-        server
-            .object_server()
-            .at(&item_path, item.clone())
-            .await
-            .unwrap();
-
-        // Calling the interface implementation directly avoids requiring a session bus while
-        // exercising the same public Item.Delete handler with real zbus dependencies.
-        item.delete(&server, &server.object_server()).await.unwrap();
-
-        assert!(!secret_file.exists());
-        assert!(store
-            .search_collection(collection_id.clone(), Arc::new(attributes.clone()))
-            .await
-            .unwrap()
-            .is_empty());
-
-        drop(item);
-        drop(store);
-        drop(server);
-        drop(client);
-
-        let attributes_db = password_store_dir
-            .join(PASS_SUBDIR)
-            .join(&*collection_id)
-            .join("attributes.redb");
-        let db = Database::open(attributes_db).unwrap();
-        let tx = db.begin_read().unwrap();
-        let labels = tx.open_table(LABELS_TABLE).unwrap();
-        let attributes_reverse = tx.open_table(ATTRIBUTES_TABLE_REVERSE).unwrap();
-        let attributes_table = tx.open_multimap_table(ATTRIBUTES_TABLE).unwrap();
-
-        assert!(ReadableTable::get(&labels, secret_id.as_str())
-            .unwrap()
-            .is_none());
-        assert!(ReadableTable::get(&attributes_reverse, secret_id.as_str())
-            .unwrap()
-            .is_none());
-        for (key, value) in &attributes {
-            assert!(
-                !ReadableMultimapTable::get(&attributes_table, (key.as_str(), value.as_str()),)
-                    .unwrap()
-                    .map(|entry| entry.unwrap().value() == secret_id.as_str())
-                    .any(|matches| matches)
-            );
-        }
-    }
-}
